pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pwninit"
version = "0.0.1"
authors = [
{ name="0xb0tm4n", email="0xb0tm4n@proton.me" },
]
description = "A pwninit script"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "pwntools @ git+https://github.com/Gallopsled/pwntools.git@dev",
  "argparse",
  "docker",
  "mako"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pwninit = ["templates/*"]

[project.scripts]
pwninit = "pwninit.pwninit:cli"
run = "pwninit.run:cli"

[project.urls]
Homepage = "https://github.com/0xb0tm4n/pwninit.py"
Issues = "https://github.com/0xb0tm4n/pwninit.py"
